- As discussed, no real need to keep the PIN page object as a subfolder. - Moved all the PIN login into Emailer's page object directory. - Pep8 and other class name/file name cleanup.
