+++ selenium_tests/lti_emailer/Emailer_Permission_Test.py
@@ -5,7 +5,7 @@ from selenium.common.exceptions import TimeoutException
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium_tests.base_test_case import get_xl_data
 from selenium_tests.lti_emailer.Emailer_Base_Test_Case import EmailerBaseTestCase
-from selenium_tests.lti_emailer.page_objects.permission_service import PermissionServiceLocators, EmailerPermissionService
+from selenium_tests.lti_emailer.page_objects.permission_service import PermissionServiceLocators, PermissionService
 from selenium_tests.lti_emailer.Emailer_Base_Test_Case import CANVAS_ADD_USERS
 
 
@@ -17,7 +17,7 @@ class EmailerPermissionTest(EmailerBaseTestCase):
     def test_roles_access(self, user_id, given_access):
         """This test masquerades as users in roles in /test_data file defined by variable CANVAS_ADD_USERS"""
 
-        permission = EmailerPermissionService(self.driver)  # instantiate
+        permission = PermissionService(self.driver)  # instantiate
         # Loop over the list of roles that should be denied access to the Emailer Tool
         if given_access == 'no':
             # Get masquerade URL for Canvas user and masquerade as user
@@ -28,11 +28,11 @@ class EmailerPermissionTest(EmailerBaseTestCase):
             permission.get_emailer_url()
             permission.switch_frame()
             print PermissionServiceLocators._validate_denied_msg + str(user_id)
-            self.assertEqual("Unauthorized", permission.unauth_message(),
+            self.assertEqual("Unauthorized", permission.check_for_denied_access(),
                              "Error: Did not get unauth error for" + str(user_id))
 
         elif given_access == 'yes':
-            permission = EmailerPermissionService(self.driver)  # instantiate
+            permission = PermissionService(self.driver)  # instantiate
             # Get masqueraded URL for Canvas user and masquerade as user
             url = PermissionServiceLocators._masquerade_url_base + str(user_id) + "/masquerade"
             self.driver.get(url)
@@ -47,7 +47,7 @@ class EmailerPermissionTest(EmailerBaseTestCase):
                                                      .is_displayed())
             except TimeoutException:
                 return False
-            self.assertEqual(PermissionServiceLocators._mailing_list_name, permission.auth_page(),
+            self.assertEqual(PermissionServiceLocators._mailing_list_name, permission.check_for_granted_access(),
                              "Error: Cannot verify Emailer Tool for: " + str(user_id))
 
 if __name__ == "__main__":
